+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # GoSysMess Changelog
 Все изменения библиотеки GoSysMess будут документироваться на этой странице.
 
+## 2023-09-10
+### Changed
+- FactoryDataContainer -> ErrFactoryInternalNoticeDataContainer
+- FactoryInternal* -> ErrFactoryInternal*
+
 ## 2023-09-03
 ### Changed
 - ErrorId -> string